pyproject.toml
[tool.poetry]
name = "tp-machine-learning-server"
version = "0.1.0"
description = "Le TP de machine learning"
authors = ["DidiStar <rurudiamant@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
djangorestframework = "^3.16.0"
djoser = "^2.3.1"
django-cors-headers = "^4.7.0"
joblib = "^1.4.2"
numpy = "^2.2.5"
scikit-learn = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
